requirements.txt
fastapi
uvicorn
python-multipart
requests
google-cloud-vision
